backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
cryptography==44.0.2
exceptiongroup==1.2.2
fastavro==1.10.0
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypdfium2==4.30.1
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
